Include dependency issue. Should be globally refactored one day. refs #70
